Removed lighting
Since the shader was that slow on my computer, that it didn't make fun
to watch the camera-transitions, the most resource-intensive-code was
removed -- the lighting. Since normals had to be calculated, some
additional rays had to be cast. To avoid this overhead, the
fragment-color is calculated from the number of steps needed to find a
point belonging to the fractal now, what's looking pretty nice, too.
This also has the nice benefit, that the shader cound be heavily
simplified.