Music renders in own thread now
Spawning a thread to render the sound allows to omit the pre-calculation
time. This, in theory, racy -- if the render-thread doesn't get enough
CPU-time, some would mute, since the sound buffer resists in .bss and
zeros would be copied to sound-device.
But since sound playback is done in a fixed time, it's almost impossible
this happens, therefore it should be fine.