Brute-forcing optimal section-ordering
Added a python script trying all possible orderings of sections in the
final binary, compiling and linking it, and comparing the sizes.
Due to the compression, the order can make a huge difference in size.
While it takes hours for the script to run, this commit includes the
currently best order.