Added console
The former textrender.h is renamed to console.h and was changed in some
ways.
Instead of providing a generic text-rendering-API, it resembles a
terminal now. It is able to print output and to print input, that is entered
by a fictive user, and therefore has a small delay between characters,
to simulate that keys are pressed to have the text displayed.