Use only one uniform
Instead of passing the orientation-, position- and
fractal-parameters-uniforms to the GPU one by one, they are an array of
vectors now. This saves us some calls and OpenGL-functions.

Due to the way the orientation-matrix has to be constructed in the
shader, the matices in keypoints.h had to be transposed. Otherwise, it
would have been required to do this in shader, introducing additional
code.