Optimized shader, changed text-color
Some work on the shader, making it smaller, on cost of readability.
This commit also includes a change of the color of the text. This is
done in the same commit, since we're at a point where the compression is
making the impact of the executable-size of any change to the shaders
completly unpredictable.